+++ lib/components/component-toggle/component-toggle.jsx
@@ -134,6 +134,7 @@ export const ComponentToggle = (props) => {
 					<AnimatedVisibility
 						visible={useComponent}
 						className={contentClassName}
+						noInitial
 					>
 						{children}
 					</AnimatedVisibility>